refactor: merge modules
