--- src/util/filter.ts
@@ -1,3 +1,0 @@
-export function notEmpty<TValue>(value: TValue | null | undefined): value is TValue {
-  return value !== null && value !== undefined;
-}
