+++ src/util/type-check.ts
@@ -1,3 +1,7 @@
 export function unreachable(): never {
   throw new Error('unreachable code');
 }
+
+export function notEmpty<TValue>(value: TValue | null | undefined): value is TValue {
+  return value !== null && value !== undefined;
+}
